requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
google-cloud-aiplatform==1.38.0
google-cloud-bigquery==3.14.0
google-cloud-translate==3.12.0
google-cloud-firestore==2.13.1
firebase-admin==6.3.0
google-generativeai==0.3.1
tensorflow==2.15.0
tensorflow-lite==2.15.0
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
pydantic==2.5.0
pyjwt==2.8.0
bcrypt==4.1.2
requests==2.31.0
aiohttp==3.9.1
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
pylint==3.0.3
mypy==1.7.1


